+++ data/colors/themes/dark_dimmed.ts
@@ -1,4 +1,4 @@
-import {get, merge} from '../../../src/utils'
+import {get, merge} from '../../../src/utils-v1'
 import dark from './dark'
 
 const scale = {
